Split up fretboard dev page into multiple files
