+++ frontend/src/pages/fretboard_dev/fretboard_dev.rs
@@ -1,0 +1,8 @@
+use leptos::prelude::*;
+
+use crate::{ pages::fretboard_dev::shared_model_demo::SharedModelDemo};
+
+#[component]
+pub fn FretboardDevPage() -> impl IntoView {
+  view! { <SharedModelDemo /> }
+}
