+++ frontend/src/pages/fretboard_dev/mod.rs
@@ -1,0 +1,6 @@
+pub mod fretboard_dev;
+pub mod frets_editor;
+pub mod helper;
+pub mod shared_model_demo;
+
+pub use fretboard_dev::FretboardDevPage;
